{{ cookiecutter.project_slug }}/requirements.txt
click==8.1.3
{%- if cookiecutter.use_requests == "Yes" %}
requests==2.28.2
{%- endif %}
{%- if cookiecutter.use_requests_cache == "Yes" %}
requests-cache==0.9.8
{%- endif %}
python-dotenv==0.21.1
pandas==2.0.0
tqdm==4.65.0
